Updated Examples with Adaptive Templates example 2 (work in progress) Added Debbing example
